--- maria/arrays/act.yml
@@ -1,79 +1,0 @@
-AtLAST:
-  description: Atacama Large-Aperture Submillimeter Telescope
-  detectors:
-    27GHz: [27.e+9, 5.e+9, 1000]
-    39GHz: [39.e+9, 5.e+9, 1000]
-    93GHz: [93.e+9, 10.e+9, 1000]
-    145GHz: [145.e+9, 10.e+9, 1000]
-    225GHz: [225.e+9, 30.e+9, 1000]
-    280GHz: [280.e+9, 40.e+9, 1000]
-  geometry: hex
-  field_of_view: 1.3
-  primary_size: 50
-  band_grouping: randomized
-  az_bounds: [0,  360]
-  el_bounds: [20,  90]
-  max_az_vel: 3
-  max_el_vel: 2
-  max_az_acc: 1
-  max_el_acc: 0.25
-  
-ACT_PA4:
-  description: ACT's PA4 (150/220 GHz)
-  detectors:
-    150GHz: [150.e+9, 10.e+9, 750]
-    220GHz: [220.e+9, 10.e+9, 750]
-  geometry: hex
-  field_of_view: 0.8
-  primary_size: 6
-  band_grouping: randomized
-  az_bounds: [0,  360]
-  el_bounds: [20,  90]
-  max_az_vel: 3
-  max_el_vel: 2
-  max_az_acc: 1
-  max_el_acc: 0.25
-ACT_PA7:
-  description: ACT's PA7 (27/39 GHz)
-  detectors:
-    27GHz: [27.e+9, 10.e+9, 75]
-    39GHz: [39.e+9, 10.e+9, 75]
-  geometry: hex
-  field_of_view: 0.8
-  primary_size: 6
-  band_grouping: randomized
-  az_bounds: [0,  360]
-  el_bounds: [20,  90]
-  max_az_vel: 3
-  max_el_vel: 2
-  max_az_acc: 1
-  max_el_acc: 0.25
-MUSTANG-2:
-  description: ''
-  detectors:
-    93GHz: [93.e+9, 10.e+9, 217]
-  geometry: hex
-  field_of_view: 0.07
-  primary_size: 100
-  band_grouping: randomized
-  az_bounds: [0,  360]
-  el_bounds: [20,  90]
-  max_az_vel: 3
-  max_el_vel: 2
-  max_az_acc: 1
-  max_el_acc: 0.25
-SCUBA-2:
-  description: ''
-  detectors:
-    350GHz: [350.e+9, 10.e+9, 5120]
-    650GHz: [650.e+9, 10.e+9, 5120]
-  geometry: hex
-  field_of_view: 0.12
-  primary_size: 15
-  band_grouping: randomized
-  az_bounds: [0,  360]
-  el_bounds: [20,  90]
-  max_az_vel: 3
-  max_el_vel: 2
-  max_az_acc: 1
-  max_el_acc: 0.25
