--- maria/arrays/atlast.yml
@@ -1,19 +1,0 @@
-AtLAST:
-  description: Atacama Large-Aperture Submillimeter Telescope
-  detectors:
-    27GHz: [27.e+9, 5.e+9, 1000]
-    39GHz: [39.e+9, 5.e+9, 1000]
-    93GHz: [93.e+9, 10.e+9, 1000]
-    145GHz: [145.e+9, 10.e+9, 1000]
-    225GHz: [225.e+9, 30.e+9, 1000]
-    280GHz: [280.e+9, 40.e+9, 1000]
-  geometry: hex
-  field_of_view: 1.3
-  primary_size: 50
-  band_grouping: randomized
-  az_bounds: [0,  360]
-  el_bounds: [20,  90]
-  max_az_vel: 3
-  max_el_vel: 2
-  max_az_acc: 1
-  max_el_acc: 0.25
