+++ maria/configs/sky.yml
@@ -1,5 +1,6 @@
-Input:
-  map_center: [0, 0]
+map_sim:
+  map_frame: "ra_dec"
+  map_center: [10, 4.5]
   map_res: 0.5
   map_inbright: 1
   map_units: KRJ
