+++ maria/sim.py
@@ -1,4 +1,5 @@
 import glob, os
+from . import atmosphere
 
 import numpy as np
 from astropy.io import fits
@@ -9,54 +10,46 @@ from .array import get_array, get_array_config
 from .pointing import get_pointing, get_pointing_config
 from .site import get_site, get_site_config
 
-from . import atmosphere, cmb, noise, sky, utils
+from . import cmb, noise, sky, utils
 
 here, this_filename = os.path.split(__file__)
+
+here, this_filename = os.path.split(__file__)
+
+from .array import get_array, ARRAY_PARAMS
+from .coordinator import Coordinator
+from .pointing import get_pointing, POINTING_PARAMS
+from .site import get_site, SITE_PARAMS
+from .sky import MAP_PARAMS
+from .tod import TOD
+
+from . import utils
+
+VALID_SIM_PARAMS = ARRAY_PARAMS | POINTING_PARAMS | SITE_PARAMS | MAP_PARAMS
 
 class InvalidSimulationParameterError(Exception):
     def __init__(self, invalid_keys):
         super().__init__(f"The parameters {invalid_keys} are not valid simulation parameters!")
 
-# valid sim kwargs are a kwarg in some config file
-config_files = glob.glob(f"{here}/configs/*.yml")
-VALID_SIM_KWARGS = set()
-for fp in config_files:
-    configs = utils.read_yaml(fp)
-    for key, config in configs.items():
-        VALID_SIM_KWARGS |= set(config.keys())
-
 def _validate_kwargs(kwargs):
-    invalid_keys = [key for key in kwargs.keys() if key not in VALID_SIM_KWARGS]
+    invalid_keys = [key for key in kwargs.keys() if key not in VALID_SIM_PARAMS]
     if len(invalid_keys) > 0:
         raise InvalidSimulationParameterError(invalid_keys)
-
-
 
 class Simulation(BaseSimulation):
     """
     A simulation! This is what users should touch, primarily. 
     """
     def __init__(self, 
-                 array, 
-                 pointing, 
-                 site, 
+                 array="default", 
+                 pointing="default", 
+                 site="default", 
                  atm_model=None, 
                  map_file=None, 
                  noise_model=None, 
                  **kwargs):
 
-        _validate_kwargs(kwargs)
-
-        if isinstance(array, str):
-            array = get_array(array, **kwargs)
-
-        if isinstance(pointing, str):
-            pointing = get_pointing(pointing, **kwargs)
-
-        if isinstance(site, str):
-            site = get_site(site, **kwargs)
-
-        super().__init__(array, pointing, site)
+        super().__init__(array, pointing, site, **kwargs)
 
         self.atm_model = atm_model
         if atm_model in ["linear_angular", "LA"]:
@@ -76,10 +69,10 @@ class Simulation(BaseSimulation):
     def _run(self):
 
         if self.atm_sim is not None:
-            self.atm_sim.run()
+            self.atm_sim._run()
 
         if self.map_sim is not None:
-            self.map_sim.run()
+            self.map_sim._run()
 
         # number of bands are lost here
         self.data = np.zeros((self.array.n_dets, self.pointing.n_time))
