+++ maria/tests/test_map_sims.py
@@ -1,0 +1,12 @@
+# import pytest
+# import maria
+# from maria.atmosphere import LinearAngularSimulation
+
+
+# @pytest.mark.map
+# def test_linear_angular_model():
+
+#     sim = LinearAngularSimulation(array="MUSTANG-2", pointing="STARE_0az_90el_60s", site="GBT", atm_model="linear_angular")
+#     tod = sim.run()
+
+
