+++ maria/tests/test_mock_obs.py
@@ -11,14 +11,14 @@ def test_we_observe():
     # Mandatory minimal weither settings
     # ---------------------
     array     = 'MUSTANG-2',       # Array type
-    pointing  = 'DAISY-2deg',      # Scanning strategy 
+    pointing  = 'daisy',      # Scanning strategy 
     site      = 'GBT',             # Site
     atm_model = 'linear_angular',  # The atmospheric model, set to None if you want a noiseless observation.
     # atm_model = None,              # The atmospheric model, set to None if you want a noiseless observation.
     
     # True sky input
     # ---------------------
-    map_file     = "./maps/protocluster.fits",                     # Input files must be a fits file.
+    map_file     = "maps/protocluster.fits",                     # Input files must be a fits file.
                                                                           # map_file can also be set to None if are only interested in the noise
     map_center   = (4.5, 10),                                             # RA & Dec in degree
 
@@ -31,7 +31,7 @@ def test_we_observe():
 
     # Additional inputs:
     # ----------------------
-    quantiles    = {'column_water_vapor' : 0.5},    # Weather conditions specific for that site
+    weather_quantiles    = {'column_water_vapor' : 0.5},    # Weather conditions specific for that site
     map_units    = 'Jy/pixel',                      # Kelvin Rayleigh Jeans (K, defeault) or Jy/pixel 
     # map_inbright = -6e-6,                        # Linearly scale the map to have this peak value.
     map_res      = 0.1 / 1000,                      # degree, overwrites header information
