+++ maria/utils/units.py
@@ -1,0 +1,55 @@
+import numpy as np
+import astropy.constants as const
+from astropy import units as u
+
+# Kelvin CMB to Jy/pixel
+# ----------------------------------------------------------------------
+global Tcmb
+Tcmb = 2.7255
+
+
+def getJynorm():
+    factor = 2e26
+    factor *= (const.k_B * Tcmb * u.Kelvin) ** 3  # (kboltz*Tcmb)**3.0
+    factor /= (const.h * const.c) ** 2  # (hplanck*clight)**2.0
+    return factor.value
+
+
+def getx(freq):
+    factor = const.h * freq * u.Hz / const.k_B / (Tcmb * u.Kelvin)
+    return factor.to(u.dimensionless_unscaled).value
+
+
+def KcmbToJyPix(freq, ipix, jpix):
+    x = getx(freq)
+    factor = getJynorm() / Tcmb
+    factor *= (x**4) * np.exp(x) / (np.expm1(x) ** 2)
+    factor *= np.abs(ipix * jpix) * (np.pi / 1.8e2) * (np.pi / 1.8e2)
+    return factor
+
+
+def KcmbToJy(freq):
+    x = getx(freq)
+    factor = getJynorm() / Tcmb
+    factor *= (x**4) * np.exp(x) / (np.expm1(x) ** 2)
+    return factor
+
+# Kelvin CMB to Kelvin brightness
+# ----------------------------------------------------------------------
+def KcmbToKbright(freq): 
+  x = getx(freq)
+  return np.exp(x)*((x/np.expm1(x))**2)
+
+# Kelvin brightness to Jy/pixel
+# ----------------------------------------------------------------------
+def KbrightToJyPix(freq,ipix,jpix):
+  return KcmbToJyPix(freq,ipix,jpix)/KcmbToKbright(freq)
+
+# Kelvin CMB to Jy/pixel
+# ----------------------------------------------------------------------
+def KcmbToJyPix(freq,ipix,jpix):
+  x = getx(freq)
+  factor  = getJynorm()/Tcmb
+  factor *= (x**4)*np.exp(x)/(np.expm1(x)**2)
+  factor *= np.abs(ipix*jpix)*(np.pi/1.8e2)*(np.pi/1.8e2)
+  return factor
